requirements_fixed.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
django-extensions==3.2.3
celery==5.3.1
redis==4.6.0
django-filter==23.2
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
django-debug-toolbar==4.2.0
openai==1.12.0
MetaTrader5==5.0.5120
kafka-python==2.2.15
pytz==2024.1
tzdata==2024.1